[TASK] change ID and JAN to i64 since SQLite doesn't support the previously used u128

+++ src/cli.rs
@@ -42,7 +42,7 @@ pub(crate) enum AddSubCommand {
 #[derive(Clap)]
 pub(crate) struct AddItem {
     /// JAN numbers of the items to add to the tracked items
-    pub(crate) input: Vec<u128>,
+    pub(crate) input: Vec<i64>,
 }
 
 /// Add an account to the database
